@@ -29,8 +29,8 @@
     "jstransform": "^11.0.0"
   },
   "devDependencies": {
-    "babel": "^5.3.3",
-    "babel-eslint": "3.1.9",
+    "babel": "^5.5.5",
+    "babel-eslint": "^3.1.14",
     "benchmark": "~1.0.0",
     "browserify": "^9.0.3",
     "bundle-collapser": "^1.1.1",
@@ -39,7 +39,7 @@
     "derequire": "^2.0.0",
     "envify": "^3.0.0",
     "es5-shim": "^4.0.0",
-    "eslint": "^0.21.2",
+    "eslint": "^0.22.1",
     "eslint-plugin-react": "^2.5.0",
     "eslint-tester": "^0.7.0",
     "grunt": "~0.4.2",
